[Refactor] Move exceptiongroup to new folder to support skip for syntax error
Closes: ray-project/ray#45702
