+++ python/ray/tests/future_python_syntax/conftest.py
@@ -1,0 +1,6 @@
+import sys
+
+collect_ignore_glob = []
+
+if sys.version_info < (3, 11):
+    collect_ignore_glob.append("*_py311.py")
